Make settings parsing fully generic using value lookup maps
Did away with the setting enums in favor of maps for easier lookup
and handling. This adds the benefit of a friendlier user interface.
It also allowed of reducing the amount of code necessary to parse
and update camera settings.

Also turns out the 0xd028 does not refer to the shutter type (MS/ES)
but to whether any of aperture/shutter-speed have hit a min/max limit.